* 4633_cleanup: (47 commits)
  Update po/*.po files.
  (tar_short_read): fix undefined behaviour.
  Fix usage of g_file_test().
  (do_link): fix typo in error message string
  diffviewer: refactoring.
  (edit_draw_this_line): use different variable for char attributes.
  tar (keyword_item_run): fix restoring extended attributes from global PAX headers.
  filemanager/chown: refactoring.
  filemanager/chattr: refactoring.
  filemanager/achown: refactoring.
  filemanager/chmod: refactoring.
  (init_learn): use button_get_width() to calculate button x-position.
  Rename variables and struct members that store return value of button_get_width().
  (button_get_width): rename from button_get_len().
  (query_dialog): refactoring.
  src/main.c: remove unused includes.
  (codepage_desc): move struct definition to c-file.
  (get_codepage_name): new function.
  Use get_codepage_id() instead of direct access to codepage_desc::id.
  lib/charsets.[ch]: cleanup.
  ...

+++ lib/tty/key.c
@@ -1279,7 +1279,7 @@ lookup_keyname (const char *name, int *idx)
 /* --------------------------------------------------------------------------------------------- */
 
 static gboolean
-lookup_keycode (const long code, int *idx)
+lookup_keycode (const int code, int *idx)
 {
     if (code != 0)
     {
